Added new system for builds for easier build control
