+++ games/Racing/project.godot
@@ -11,9 +11,11 @@ config_version=5
 [application]
 
 config/name="Racing"
-run/main_scene="res://scenes/train.tscn"
+run/main_scene="res://scenes/game.tscn"
 config/features=PackedStringArray("4.1", "Forward Plus")
 config/icon="res://icon.svg"
+run/main_scene.test="res://scenes/test.tscn"
+run/main_scene.train="res://scenes/train.tscn"
 
 [display]
