+++ scripts/test/test_racing.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 python scripts/sb3_test.py \
-    --env_path="games/Racing/builds/racing.x86_64" \
-    --agents_per_env=6 \
+    --env_path="games/Racing/builds/racing_test.x86_64" \
+    --agents_per_env=4 \
     --action_repeat=8 \
     --load_model_paths="models/Racing/Racing_0/Racing_0, models/Racing/Racing_0/Racing_0" \
     --num_episodes=5 \
